dpif-netdev: e2e: Calculate ct_counters based on pointers values
In e2e_cache_calc_counters() calculate ct_counter of merged flow
a minimum value of pointers to MT flows objects. Guarantees uniqueness
and eliminates hash calculation. Propagate uintptr_t value to
offload layer.

Issue: 2333202

+++ lib/netdev-offload-dpdk.c
@@ -2258,7 +2258,7 @@ struct act_vars {
     struct rte_flow_action_jump *jump;
     bool is_e2e_cache_flow;
     uint32_t app_flows_counter;
-    uint32_t app_ct_counter;
+    uintptr_t app_ct_counter;
     struct rte_flow_action *shared;
 };
